fix: correct E/Z label font settings access and clean up minor anti-patterns
bond_item.py: E/Z label in paint() fetched font_size/family via
win.settings.get() (MainWindow direct attr), but settings live on
init_manager. hasattr(win, 'settings') was always False, so E/Z
labels ignored user font settings. Now uses scene.get_setting() like
the rest of bond_item.py.

mol_geometry.py: identify_valence_problems() accumulated bond orders
into Dict[int, int] but aromatic bonds (order=1.5) make values float;
corrected annotation to Dict[int, float].

plugin_interface.py: to_xyz_block() bare except swallowed errors
silently; now logs at DEBUG level before returning None.

app_state.py: redundant try/except wrapped around contextlib.suppress
catching the same exception types; removed the outer try/except.

+++ moleditpy/src/moleditpy/core/mol_geometry.py
@@ -504,7 +504,7 @@ def identify_valence_problems(
     problem_atom_ids = []
 
     # Pre-calculate bond orders per atom
-    bond_orders: Dict[int, int] = {}
+    bond_orders: Dict[int, float] = {}
     for (id1, id2), bond in bonds_data.items():
         order = bond.get("order", 1)
         bond_orders[id1] = bond_orders.get(id1, 0) + order
